+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  - Ignore BCC errors in select iso1443a tags process
  - Mikron JSC Russia Ultralight EV1 41 pages tag type support
  - Add test for Ultralight gen2 magic 'hf search'
+ - Add test for Ultralight EV1 gen2 magic 'hf search'
 
 # Change Log
 All notable changes to this project will be documented in this file.